src/test/java/org/springframework/beans/factory/ToykoMetroKanjiHiraganaMapFactoryBeanTest.java
package org.springframework.beans.factory;

import java.lang.reflect.InvocationTargetException;
import java.lang.reflect.Method;
import java.util.Map;
import java.util.function.Predicate;

import org.apache.commons.lang3.function.FailableFunction;
import org.apache.commons.lang3.reflect.FieldUtils;
import org.jsoup.nodes.TextNode;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.meeuw.functional.Predicates;

import io.github.toolfactory.narcissus.Narcissus;

class ToykoMetroKanjiHiraganaMapFactoryBeanTest {

	private static Method METHOD_TEXT_TEXT_NODE, METHOD_TEST_AND_APPLY = null;

	@BeforeAll
	static void beforeAll() throws ReflectiveOperationException {
		//
		final Class<?> clz = ToykoMetroKanjiHiraganaMapFactoryBean.class;
		//
		(METHOD_TEXT_TEXT_NODE = clz.getDeclaredMethod("text", TextNode.class)).setAccessible(true);
		//
		(METHOD_TEST_AND_APPLY = clz.getDeclaredMethod("testAndApply", Predicate.class, Object.class,
				FailableFunction.class, FailableFunction.class)).setAccessible(true);
		//
	}

	private ToykoMetroKanjiHiraganaMapFactoryBean instance = null;

	@BeforeEach
	void beforeEach() {
		//
		instance = new ToykoMetroKanjiHiraganaMapFactoryBean();
		//
	}

	@Test
	void testGetObject() throws Exception {
		//
		Assertions.assertNull(instance != null ? instance.getObject() : null);
		//
	}

	@Test
	void testGetObjectType() {
		//
		Assertions.assertEquals(Map.class, instance != null ? instance.getObjectType() : null);
		//
	}

	@Test
	void testText() throws Throwable {
		//
		// org.jsoup.nodes.TextNode
		//
		Assertions.assertNull(text((TextNode) null));
		//
		final TextNode textNode = Util.cast(TextNode.class, Narcissus.allocateInstance(TextNode.class));
		//
		Assertions.assertNull(text(textNode));
		//
		final String string = "";
		//
		FieldUtils.writeField(textNode, "value", string, true);
		//
		Assertions.assertSame(string, text(textNode));
		//
	}

	private static String text(final TextNode instnace) throws Throwable {
		try {
			final Object obj = METHOD_TEXT_TEXT_NODE.invoke(null, instnace);
			if (obj == null) {
				return null;
			} else if (obj instanceof String) {
				return (String) obj;
			}
			throw new Throwable(obj.getClass() != null ? obj.getClass().toString() : null);
		} catch (final InvocationTargetException e) {
			throw e.getTargetException();
		}
	}

	@Test
	void testTestAndApply() throws Throwable {
		//
		Assertions.assertNull(testAndApply(null, null, null, null));
		//
		Assertions.assertNull(testAndApply(Predicates.alwaysTrue(), null, null, null));
		//
	}

	private static <T, R, E extends Throwable> R testAndApply(final Predicate<T> predicate, final T value,
			final FailableFunction<T, R, E> functionTrue, final FailableFunction<T, R, E> functionFalse)
			throws Throwable {
		try {
			return (R) METHOD_TEST_AND_APPLY.invoke(null, predicate, value, functionTrue, functionFalse);
		} catch (final InvocationTargetException e) {
			throw e.getTargetException();
		}
	}

}
